+++ backend/src/models/reports/genericModelTemplate.js
@@ -1,36 +1,24 @@
 /**
- * A GENERIC MODEL STARTER TEMPLATE FOR FUTURE REPORTs
+ * A GENERIC MODEL STARTER TEMPLATE FOR FUTURE REPORT
+ *
+ * This should generate a 'fiscal year' table in the backend console out-of-the box
+ * THIS HELPS WITH STEP 2 of building a report: getting the backend to generate a JSON object and blank .excel document
  * you can adapt this to help structure your model.
  * this is based on Tab_20_rpt_PA_Billed.js AND Tab_50_rpt_PF_NetRecoverySummaryByQuarter.js:w
+ *
+ * This template tries to just use the raw query, but look at tab 20 or tab 21 if you want
+ * to wrap it further in knex() methods.
  *
  */
 // libs
 const { knex } = require("@database/databaseConnection")();
-
-// Constants
-// eslint-disable-next-line no-unused-vars
-const PARAMETER = "someParameter";
-
-/**
- * Base query to re-use for report & report totals
- *
- * @param   {number | string | Array} parameter from the frontend, e.g.: Date, Portfolio(s), fiscal_year, or the like
- * @returns {Knex.QueryBuilder}                 Knex query builder for fetching report totals.
- */
-const baseQuery = knex("<tableName>")
-  .select({
-    someColumn: "<some value>",
-  })
-  .leftJoin("<some table and criteria>")
-  .groupBy("<some criteria>")
-  .orderBy("P.project_number", "asc");
 
 /**
  * Retrieves the data for various financial metrics based on the fiscal year.
  *
  * Uses baseQuery twice, for DRYness
  *
- * @param   {number | string | Array} Parameter- The fiscal, Date, Portfolio(s) to grab data for
+ * @param   {number | string | Array} Parameter- The fiscal, Date, or Portfolio(s) to grab data for
  * @returns {Promise}                            - A promise that resolves to the query result
  */
 const reportQueries = {
@@ -38,21 +26,62 @@ const reportQueries = {
     knex("fiscal_year").select("fiscal_year").where("fiscal_year.id", PARAMETER).first(),
 
   report: (PARAMETER) =>
-    knex.select("something").from(baseQuery.as("base")).where("someColumn", PARAMETER),
+    knex
+      .select()
+      .fromRaw(
+        `
+        (
 
-  report_totals: (PARAMETER) =>
-    knex.from(baseQuery.as("base")).sum("someColumn").where("someColumn", PARAMETER),
+
+
+        PASTE RAW QUERY HERE
+
+
+
+
+        ) as base`
+      )
+      .where("fiscal", PARAMETER),
+
+  totals: (PARAMETER) =>
+    knex
+      .select()
+      .fromRaw(
+        `
+        (
+
+
+
+        PASTE RAW QUERY HERE
+
+
+
+
+        ) as base`
+      )
+      .where("fiscal", PARAMETER),
 };
 
 module.exports = {
-  required: ["PARAMETER"], // e.g. fiscal, date, or portfolio
-  getAll: async ({ PARAMETER }) => {
-    const [{ fiscal_year }, report, report_totals] = await Promise.all([
+  required: ["fiscal"], // e.g. fiscal, date, or portfolio
+  getAll: async ({ fiscal: PARAMETER }) => {
+    // replace fiscal  above with whatever parameter you take here
+    const [{ fiscal_year } /*report, report_totals*/] = await Promise.all([
       reportQueries.fiscalYear(PARAMETER),
+      /*
       reportQueries.report(PARAMETER),
-      reportQueries.report_totals(PARAMETER),
+      reportQueries.totals(PARAMETER),
+      */
     ]);
 
-    return { fiscal_year, report, report_totals };
+    const reportData = { fiscal_year /*report, totals*/ };
+
+    /**
+     * console.warn doesn't produce a linter error
+     * delete after your template is populating.
+     */
+    console.warn(JSON.stringify(reportData, null, 2));
+
+    return reportData;
   },
 };